requirements.txt
Django==3.2.4
gunicorn
djangorestframework
psycopg2


# Install dependencies
pytorch-cuda == 11.8
mediapipe
opencv-python
numpy
pandas
torchinfo
keras